+++ packages/orq-rc/RELEASES.md
@@ -1479,3 +1479,13 @@ Based on:
 - [typescript v3.4.0-rc.35] packages/orq-rc
 ### Releases
 - [NPM v3.4.0-rc.35] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.35 - packages/orq-rc
+
+## 2025-04-17 14:28:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.533.0 (2.578.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.4.0-rc.36] packages/orq-rc
+### Releases
+- [NPM v3.4.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/3.4.0-rc.36 - packages/orq-rc